+++ weather_station/measurements_devices/veml6075.py
@@ -2,8 +2,6 @@ import board
 import busio
 import adafruit_veml6075
 
-from decimal import Decimal
-
 i2c = busio.I2C(board.SCL, board.SDA)
 
 veml = adafruit_veml6075.VEML6075(i2c, integration_time=100)
@@ -11,7 +9,7 @@ veml = adafruit_veml6075.VEML6075(i2c, integration_time=100)
 
 def get_data():
     return {
-        "uv_index": Decimal(str(round(veml.uv_index, 2))),
-        "uv_a": Decimal(str(round(veml.uva, 1))),
-        "uv_b": Decimal(str(round(veml.uvb, 1))),
+        "uv_index": str(round(veml.uv_index, 2)),
+        "uv_a": str(round(veml.uva, 1)),
+        "uv_b": str(round(veml.uvb, 1)),
     }
